+++ CHANGELOG.md
@@ -29,8 +29,12 @@
 * `SBBHeader`: Renamed field `sbbSignetTooltip` to `logoTooltip`
 * `SBBOnboarding`: Padding now defined by `SBBThemeData.defaultRootContainerPadding` 
 * `SBBTextStyles`: Adjusted `fontSize` and `height` values to match the current specifications
+* `SBBListHeader`: Adjusted paddings to match the current specifications
+* `SBBTextField`: Adjusted paddings to match the current specifications
 ### Deprecated
 ### Removed
+* `SBBListHeader`: Removed fields `icon` and `onCallToAction` to match the current specifications
+* `SBBTextField`: Removed field `alignLabelWithHint`
 ### Fixed
 * `SBBHeader`: SBB Logo now excluded from semantics if `onPressedLogo` is `null`
 * `SBBModalSheet`: Fine line that sometimes was visible below the header is now gone